+++ client/src/Components/Footer/Footer.styled.js
@@ -1,132 +1,7 @@
-import React from "react";
 import styled from "styled-components";
-import { Link } from "react-router-dom";
-import {
-	LazyLoadImage,
-	trackWindowScroll,
-} from "react-lazy-load-image-component";
-import "react-lazy-load-image-component/src/effects/blur.css";
-
-function Footer({ scrollPosition }) {
-	return (
-		<Container>
-			<Content>
-				<Social>
-					<div>
-						<a href="/" className="logo">
-							<LazyLoadImage
-								className="gallery-image"
-								effect="blur"
-								width="40px"
-								height="40px"
-								src="/images/logo.svg"
-								alt="Logo"
-								scrollPosition={scrollPosition}
-							/>
-						</a>
-
-						<p>Collection of interesting and funny programming memes.</p>
-					</div>
-
-					<Icons>
-						<p>Follow us on all socials</p>
-						<div>
-							<a
-								href="https://web.facebook.com/evavic44"
-								title="Facebook"
-								aria-label="Facebook"
-							>
-								<i className="fab fa-facebook"></i>
-							</a>
-							<a
-								href="https://twitter.com/evavic44"
-								alt="Twitter icon"
-								title="Twitter"
-								aria-label="Twitter"
-							>
-								<i className="fab fa-twitter"></i>
-							</a>
-							<a
-								href="https://www.linkedin.com/in/evavic44"
-								title="LinkedIn"
-								aria-label="LinkedIn"
-							>
-								<i className="fab fa-linkedin"></i>
-							</a>
-							<a
-								href="https://github.com/Evavic44/rocketmeme"
-								title="GitHub"
-								aria-label="GitHub"
-							>
-								<i className="fab fa-github"></i>
-							</a>
-							<a
-								href="https://eke.hashnode.dev/rocketmeme-a-meme-library-and-editor-for-creating-and-sharing-memes"
-								title="Hashnode"
-								aria-label="Hashnode"
-							>
-								<i className="fab fa-hashnode"></i>
-							</a>
-						</div>
-					</Icons>
-				</Social>
-
-				{/* Footer links */}
-				<FooterLink>
-					<div className="footer-categories">
-						<h3>Categories</h3>
-						<Link to="/categories">JavaScript</Link>
-						<Link to="/categories">Programming</Link>
-						<Link to="/categories">Stackoverflow</Link>
-						<Link to="/categories">NFT</Link>
-					</div>
-
-					<div className="footer-resources">
-						<h3>Resources</h3>
-						<a href="https://github.com/Evavic44/rocketmeme/blob/main/LICENSE">
-							License
-						</a>
-						<a href="https://github.com/Evavic44/rocketmeme/blob/main/CODE_OF_CONDUCT.md">
-							Code of Conduct
-						</a>
-						<a href="https://github.com/Evavic44/rocketmeme/blob/main/README.md">
-							Docs
-						</a>
-					</div>
-
-					<div className="footer-about">
-						<h3>About Us</h3>
-						<a href="/about">About</a>
-						<a href="mailto:evavic44@gmail.com">Contact Us</a>
-					</div>
-
-					<div className="footer-donate">
-						<h3>Donate</h3>
-						<a href="https://www.buymeacoffee.com/evavic44">Support our work</a>
-						<a href="/">Share</a>
-					</div>
-
-					<div className="footer-contribute">
-						<h3>Developer</h3>
-						<a href="https://github.com/Evavic44/rocketmeme">Docs</a>
-						<a href="https://rocketteam.gitbook.io/rocketmemes-api/">API</a>
-					</div>
-				</FooterLink>
-			</Content>
-
-			<Masthead>
-				<small>
-					Built with <i className="fas fa-heart highlight"></i> by{" "}
-					<a href="https://github.com/Evavic44">Eke</a> and{" "}
-					<a href="https://github.com/SpiffGreen">Spiff</a>
-				</small>
-			</Masthead>
-		</Container>
-	);
-}
 
 // Footer
-const Container = styled.footer`
+export const Container = styled.footer`
 	min-height: 400px;
 	background-color: var(--bg-dark);
 	color: var(--text-light);
@@ -147,7 +22,7 @@ const Container = styled.footer`
 `;
 
 // Container width
-const Content = styled.div`
+export const Content = styled.div`
 	display: flex;
 	flex-direction: row;
 	justify-content: space-around;
@@ -159,7 +34,7 @@ const Content = styled.div`
 `;
 
 // Social container
-const Social = styled.div`
+export const Social = styled.div`
 	display: flex;
 	flex-direction: column;
 	align-items: flex-start;
@@ -199,7 +74,7 @@ const Social = styled.div`
 `;
 
 // Social icons
-const Icons = styled.div`
+export const Icons = styled.div`
 	display: flex;
 	flex-direction: column;
 	font-size: 1.3rem;
@@ -228,7 +103,7 @@ const Icons = styled.div`
 `;
 
 // Footer links
-const FooterLink = styled.div`
+export const FooterLink = styled.div`
 	display: grid;
 	grid-template-columns: repeat(5, 1fr);
 
@@ -286,7 +161,7 @@ const FooterLink = styled.div`
 `;
 
 // Masthead
-const Masthead = styled.div`
+export const Masthead = styled.div`
 	display: flex;
 	flex-direction: column;
 	align-items: center;
@@ -307,5 +182,3 @@ const Masthead = styled.div`
 		margin-bottom: 0.7rem;
 	}
 `;
-
-export default Footer;
